% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_Monthly_Temperature.R
\name{Average_Monthly_Temperature}
\alias{Average_Monthly_Temperature}
\title{average temperature}
\usage{
Average_Monthly_Temperature(climate_data)
}
\description{
average temperature
}
\author{
Rae Fuhrman
}
