% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plant_Growth.R
\name{plant_growth}
\alias{plant_growth}
\title{Plant Growth}
\usage{
plant_growth(sun = sunshine_hours)
}
\arguments{
\item{Sun.}{A dataframe, multidimensional array, or other input that contains sunshine hours.  The default is sunshine_hours, a multidimensional array with sunshine hours by months and years for three different locations.}
}
\value{
The plant growth of an imaginary plant based on hours of sunshine.  Assume this growth is in mm.
}
\description{
This function calculates plant growth for an imaginary plant that grows only when sunshine hours for a certain timeframe is bigger than one.
}
\author{
Cristina
}
