% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_growth_mda.R
\docType{data}
\name{plant_growth_mda}
\alias{plant_growth_mda}
\title{Imaginary plant growth}
\format{An object of class \code{array} of dimension 6 x 12 x 3.}
\usage{
plant_growth_mda
}
\description{
Imaginary data for monthly plant growth, in mm, by months and years between 2001 and 2006 for three locations.  Based off of sunshine_hours and the Plant_Growth function used to generate growth from sunshine hours.
}
\details{
@format A multidimensional array, with years in rows, months in columns, and locations in the third dimension.
 @source Plant_Growth function applied to sunshine_hours (random data)
 \itemize{
 \item plant growth (mm)
 }
}
\keyword{datasets}
