% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_revenue.R
\name{plant_revenue}
\alias{plant_revenue}
\title{Plant Revenue}
\usage{
plant_revenue(growth = plant_growth_mda)
}
\arguments{
\item{growth.}{A dataframe, multidimensional array, or other input that contains plant growth in mm.  The default is plant_growth_mda, a multidimensional array with plant growth by months and years for three different locations.}
}
\value{
The revenue in $ from an imaginary plant for each timeframe.  The default data in multidimensional array also returns a multidimensional array.
}
\description{
This function calculates plant revenue for a very valuable, imaginary plant based on its growth in mm, where each mm growth generates $3.75 in revenue.
}
\author{
Cristina
}
