% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation_growth.R
\name{precipitation_plant_growth}
\alias{precipitation_plant_growth}
\title{Plant growth based on precipitation}
\usage{
precipitation_plant_growth(climate_data)
}
\arguments{
\item{PRCP}{A variable in dataframe called climate_data. The default is climate_data, a dataframe with sunshine hours by months and years for one station in Santa Barbara.}
}
\value{
A list of the plant growth of an imaginary plant based on cm of precipitation for each value of precipitation associated with a date. Assume this growth is in mm.
}
\description{
This function calculates plant growth for an imaginary plant that grows only when precipitation values are above a threshold (0.5 cm)
}
\author{
Rae Fuhrman
}
