% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunshine_hours.R
\docType{data}
\name{sunshine_hours}
\alias{sunshine_hours}
\title{Imaginary sunshine hours}
\format{An object of class \code{array} of dimension 6 x 12 x 3.}
\usage{
sunshine_hours
}
\description{
Imaginary data for average daily sunshine hours by month and year (2001-2006) for three imaginary locations
}
\details{
@format A multidimensional array, with years in rows, months in columns, and locations in the third dimension.
 @source Random numbers generated between 0 and 10 for each year-month-location
 \itemize{
 \item sunshine hours
 }
}
\keyword{datasets}
