% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_precipitation.R
\name{Total_yearly_precipitation}
\alias{Total_yearly_precipitation}
\title{Total yearly precipitation}
\usage{
Total_yearly_precipitation(climate_data)
}
\value{
Total precipitation in 2018 at the station USC00047902
}
\description{
This function sums the yearly precipitation at a specific station in Santa Barbara in 2018 based on a climate dataset from NOAA.
}
\author{
Rae Fuhrman
}
