% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_precipitation.R
\name{Total_yearly_precipitation}
\alias{Total_yearly_precipitation}
\title{total yearly temperature}
\usage{
Total_yearly_precipitation(climate_data)
}
\description{
total yearly temperature
}
\author{
Rae Fuhrman
}
